fix: rename Dropzone to Upload -clean usages -add missing validator in upload button -dont export upload area, move it to elements
