+++ packages/upload/package.json
@@ -29,7 +29,7 @@
   "peerDependencies": {
     "react": "^16.11.0",
     "styled-components": "^5.0.1",
-    "woly": "^0.7.0"
+    "woly": "^0.9.0"
   },
   "scripts": {},
   "author": {
