+++ src/upload/index.ts
@@ -1,2 +1,1 @@
-export * from './atoms';
 export * from './molecules';
